requirements.txt
smolagents
litellm
python-dotenv
huggingface_hub 
flask
flask-cors